+++ Changes.md
@@ -8,6 +8,7 @@ Version 2.9.1.7
   5. Refactor FontAwesome / GlyphIcon markup generation.
   6. Fix margin on social network icons.
   7. Fix user profile image hover.
+  8. Fix: Block visibility issue with block, content, block layout - #23.
 
 Version 2.9.1.6
   1. Update to FontAwesome 4.5.